Issue #8059 : Rebrand hop-env files as Hop install spec and fix Open/Save As
Marketplace Open/Save As now uses HopVfs and variables so project-relative
paths work. Rename the Environment tab to Install spec to distinguish these
YAML files from project lifecycle environments. Add an Explorer file type
and Edit YAML in Explorer for hand edits.
