+++ plugins/misc/marketplace/src/main/java/org/apache/hop/marketplace/gui/MarketplaceDialog.java
@@ -18,7 +18,6 @@
 package org.apache.hop.marketplace.gui;
 
 import java.lang.reflect.InvocationTargetException;
-import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.LinkedHashMap;
@@ -40,6 +39,7 @@ import org.apache.hop.marketplace.command.MarketplaceCommand;
 import org.apache.hop.marketplace.config.MarketplaceConfig;
 import org.apache.hop.marketplace.config.MarketplaceRepository;
 import org.apache.hop.marketplace.env.EnvironmentApplier;
+import org.apache.hop.marketplace.env.HopInstallSpecFiles;
 import org.apache.hop.marketplace.install.HopHome;
 import org.apache.hop.marketplace.install.InstallReceipt;
 import org.apache.hop.marketplace.install.PluginInstaller;
@@ -53,6 +53,7 @@ import org.apache.hop.ui.core.gui.GuiResource;
 import org.apache.hop.ui.core.gui.WindowProperty;
 import org.apache.hop.ui.core.widget.ColumnInfo;
 import org.apache.hop.ui.core.widget.TableView;
+import org.apache.hop.ui.hopgui.HopGui;
 import org.apache.hop.ui.pipeline.transform.BaseTransformDialog;
 import org.eclipse.swt.SWT;
 import org.eclipse.swt.custom.CTabFolder;
@@ -119,14 +120,14 @@ public class MarketplaceDialog extends Dialog {
    * @param initialSearch search box content on open, or null/blank to list everything
    */
   public MarketplaceDialog(Shell parent, String initialSearch) {
-    super(parent, SWT.DIALOG_TRIM | SWT.APPLICATION_MODAL | SWT.RESIZE | SWT.MAX);
+    super(parent, SWT.DIALOG_TRIM | SWT.RESIZE | SWT.MAX);
     this.props = PropsUi.getInstance();
     this.initialSearch = initialSearch;
   }
 
   public void open() {
     Shell parent = getParent();
-    shell = new Shell(parent, SWT.DIALOG_TRIM | SWT.APPLICATION_MODAL | SWT.RESIZE | SWT.MAX);
+    shell = new Shell(parent, SWT.DIALOG_TRIM | SWT.RESIZE | SWT.MAX);
     PropsUi.setLook(shell);
     shell.setImage(GuiResource.getInstance().getImageMarketplace());
     shell.setText(BaseMessages.getString(PKG, "MarketplaceDialog.Shell.Title"));
@@ -341,8 +342,8 @@ public class MarketplaceDialog extends Dialog {
 
   private void createEnvironmentTab(CTabFolder folder) {
     CTabItem tab = new CTabItem(folder, SWT.NONE);
-    tab.setText(BaseMessages.getString(PKG, "MarketplaceDialog.Tab.Environment"));
-    tab.setImage(GuiResource.getInstance().getImageClientEnvironment());
+    tab.setText(BaseMessages.getString(PKG, "MarketplaceDialog.Tab.InstallSpec"));
+    tab.setImage(GuiResource.getInstance().getImageMarketplace());
     Composite comp = new Composite(folder, SWT.NONE);
     PropsUi.setLook(comp);
     FormLayout layout = new FormLayout();
@@ -351,8 +352,7 @@ public class MarketplaceDialog extends Dialog {
     comp.setLayout(layout);
     tab.setControl(comp);
 
-    // Full hop-env editor (formerly HopEnvironmentDialog modal)
-    HopEnvironmentDialog.embed(
+    HopInstallSpecEditor.embed(
         comp,
         resolveDefaultEnvPath(),
         msg -> {
@@ -362,22 +362,33 @@ public class MarketplaceDialog extends Dialog {
         });
   }
 
-  private Path resolveDefaultEnvPath() {
-    // Last file edited in the marketplace environment editor (audit list, most-recent first)
+  private String resolveDefaultEnvPath() {
     String last =
         org.apache.hop.ui.hopgui.shared.AuditManagerGuiUtil.getLastUsedValue(
-            HopEnvironmentDialog.AUDIT_TYPE_ENV_FILES);
-    if (StringUtils.isNotBlank(last)) {
-      Path lastPath = Path.of(last.trim()).toAbsolutePath().normalize();
-      if (Files.isRegularFile(lastPath)) {
-        return lastPath;
+            HopInstallSpecEditor.AUDIT_TYPE_ENV_FILES);
+    if (HopInstallSpecFiles.exists(last, hopGuiVariables())) {
+      return last;
+    }
+    if (hopHome != null) {
+      String fullClient = hopHome.resolve("full-client-env.yaml").toString();
+      if (HopInstallSpecFiles.exists(fullClient, hopGuiVariables())) {
+        return fullClient;
       }
     }
-    Path fullClient = hopHome.resolve("full-client-env.yaml");
-    if (Files.isRegularFile(fullClient)) {
-      return fullClient;
+    Path discovered = EnvironmentApplier.resolveEnvironmentFile(hopHome, null);
+    return discovered == null ? null : discovered.toString();
+  }
+
+  private org.apache.hop.core.variables.IVariables hopGuiVariables() {
+    try {
+      HopGui hopGui = HopGui.getInstance();
+      if (hopGui != null && hopGui.getVariables() != null) {
+        return hopGui.getVariables();
+      }
+    } catch (Exception ignored) {
+      // dialog constructed outside Hop Gui
     }
-    return EnvironmentApplier.resolveEnvironmentFile(hopHome, null);
+    return Variables.getADefaultVariableSpace();
   }
 
   private void createRepositoriesTab(CTabFolder folder) {
